requirements.txt
beautifulsoup4>=4.12.0
requests>=2.31.0
pandas>=2.0.0
selenium>=4.15.0
playwright>=1.40.0
plotly>=5.18.0
scikit-learn>=1.3.0
scipy>=1.11.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
newsapi-python>=0.2.7
streamlit>=1.28.0
